Improved Items + Added Despawn timer script (for particle systems) + Overhauled Enemy hitboxes + Minor changes to Player attack hitboxes
Improved Items + Added Despawn timer script (for particle systems) + Overhauled Enemy hitboxes + Minor changes to Player attack hitboxes

Known issue: Attack hitbox does not detect collisions when it is activated while the enemy is already touching the player. Also, the enemy is not able to hit the player in that situation. In combination this leads to a deadlock situation where no one takes hits until the player moves (not tested on any mobile device yet)